+++ integration_tests/test_cli_spatial_indexes.py
@@ -6,7 +6,7 @@
 import pytest
 
 
-@pytest.mark.parametrize('datacube_env_name', ('experimental',))
+@pytest.mark.parametrize('datacube_env_name', ('postgis',))
 def test_cli_spatial_indexes(index, clirunner):
     runner = clirunner(['spindex', 'list'], verbose_flag=False, expect_success=True)
     assert "EPSG:4326" in runner.output
@@ -48,7 +48,7 @@ def test_cli_spatial_indexes(index, clirunner):
     assert runner.exit_code == 0
 
 
-@pytest.mark.parametrize('datacube_env_name', ('experimental',))
+@pytest.mark.parametrize('datacube_env_name', ('postgis',))
 def test_cli_spatial_index_create_and_update(index, clirunner):
     runner = clirunner(['spindex', 'list'], verbose_flag=False, expect_success=True)
     assert "EPSG:4326" in runner.output
@@ -88,7 +88,7 @@ def test_cli_spatial_indexes_on_non_supporting_index(index, clirunner):
     assert runner.exit_code == 1
 
 
-@pytest.mark.parametrize('datacube_env_name', ('experimental',))
+@pytest.mark.parametrize('datacube_env_name', ('postgis',))
 def test_cli_spatial_indexes_no_srids(index, clirunner):
     runner = clirunner(['spindex', 'create'], verbose_flag=False, expect_success=False)
     assert "Must supply at least one CRS" in runner.output
@@ -103,7 +103,7 @@ def test_cli_spatial_indexes_no_srids(index, clirunner):
     assert runner.exit_code == 1
 
 
-@pytest.mark.parametrize('datacube_env_name', ('experimental',))
+@pytest.mark.parametrize('datacube_env_name', ('postgis',))
 def test_cli_spatial_indexes_bad_srid(index, clirunner):
     runner = clirunner(['spindex', 'create', '1'], verbose_flag=False, expect_success=False)
     assert runner.exit_code == 1